Adding -NoWait parameter for alerts with buttons
•Adding -NoWait parameter to launch alerts with buttons without forcing
script to wait for user to click a button. This should be most useful
for alerts that only use an "OK" button that does not need to be
tracked.
•Using Invoke-WMIMethod for -NoWait alerts, since the KACE agent tracks
(and waits on) child processes launched with *Start Process* and so
scripts would remain "Running" until the user clicked a button, even
after the powershell script had quit.